+++ backend/app/tests/users/test_user.py
@@ -23,21 +23,19 @@ def test_get_users(client, user):
     user_schema = UserPublicSchema.model_validate(user)
     response = client.get("/users")
     assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'users': [user_schema.model_dump(mode="json")]}
+    assert response.json() == {"users": [user_schema.model_dump(mode="json")]}
 
 
 def test_update_user(client, user, token):
     response = client.put(
         f"/users/{user.id}",
         json={"username": "updatedusername", "email": "newemail@test.com", "password": "newpassword"},
-        headers={"Authorization": f"Bearer {token}"}
+        headers={"Authorization": f"Bearer {token}"},
     )
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {"id": user.id, "username": "updatedusername", "email": "newemail@test.com"}
 
 
 def test_delete_user(client, user, token):
-    response = client.delete(
-        "/users/1",
-        headers={"Authorization": f"Bearer {token}"})
+    response = client.delete("/users/1", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == HTTPStatus.NO_CONTENT
